page_id_width is now a uint8_t type

+++ inc/bplus_tree_tuple_definitions.h
@@ -7,7 +7,7 @@ typedef struct bplus_tree_tuple_defs bplus_tree_tuple_defs;
 struct bplus_tree_tuple_defs
 {
 	// page_id_with is bytes required for storing page_id, it can be 1,2,4 or 8
-	int page_id_width;
+	uint8_t page_id_width;
 
 	// size of each page inside this bplus tree
 	uint32_t page_size;
@@ -39,7 +39,7 @@ struct bplus_tree_tuple_defs
 // it allocates memory only for index_def
 // returns 1 for success, it fails with 0, if the record_def has element_count 0 OR key_element_count == 0 OR key_elements == NULL OR if any of the key_element_ids is out of bounds
 // page_id_with is bytes required for storing page_id, it can be 1,2,4 or 8
-int init_bplus_tree_tuple_definitions(bplus_tree_tuple_defs* bpttd_p, const tuple_def* record_def, const uint32_t* key_element_ids, uint32_t key_element_count, uint32_t page_size, int page_id_width, uint64_t NULL_PAGE_ID);
+int init_bplus_tree_tuple_definitions(bplus_tree_tuple_defs* bpttd_p, const tuple_def* record_def, const uint32_t* key_element_ids, uint32_t key_element_count, uint32_t page_size, uint8_t page_id_width, uint64_t NULL_PAGE_ID);
 
 // get maximum size of the record that can be inserted in this bplus_tree
 // there must be atleast 2 data records per leaf page and 2 index records per interior page of any bplus_tree
